+++ Liber/Models/Book.swift
@@ -11,13 +11,25 @@ import Foundation
 // Handles the data for a book.
 class Book {
     var title : String = ""
-    var author : String = ""
-    var isbn_13 : Int = 0
-    var isbn_10 : Int = 0
-    var edition : Int = 0
-    var binding : String = ""
+    var author = [String]()
+    var isbn_13 : String = ""
+    var isbn_10 : String = ""
     var publisher : String = ""
-    var published = Date()
-    var price : Double = 0.0
-    var location : String = ""
+    var published : String = ""
+    var thumbnail = URL(string: "")
+    //var price : String = ""
+    //var location : String = ""
+    
+    
+    func resetClassData() {
+        title = ""
+        author.removeAll()
+        isbn_13 = ""
+        isbn_10 = ""
+        publisher = ""
+        published = ""
+        thumbnail = URL(string: "")
+        //price = ""
+        //location = ""
+    }
 }